Azure Functions support for Node 12 is ending on 13 December 2022
Node 12 is only supported by Azure Functions host version 3, which is ending support on 13 December 2022. Node 12 also reached the end of community support 30 April 2022. 
As such, we recommend developers to update their functions apps to use Azure Functions host version 4 and Node 16. The Functions service will continue to run applications that are running Node 12 beyond 13 December 2022, but if support is needed customers will be asked to upgrade to Node 16.